Fix gamut mask rendering
quint8 will overflow if the view size is >255 pixels.
Since all QTransform related calculations use qreal anyway,
change the function arguments to qreal.

Also, Artistic Color Selector didn't update the transform
on resize, and the outer circle of Advanced Color Selector
was 1 pixel too small on odd dimensions due to integer math.

Bug:421142
(cherry picked from commit e29a4e99ef88cdbd691915f3b87c4d59d9433b5a)

+++ libs/flake/resources/KoGamutMask.cpp
@@ -182,7 +182,7 @@ void KoGamutMask::paintStroke(QPainter &painter, bool preview)
     }
 }
 
-QTransform KoGamutMask::maskToViewTransform(quint8 viewSize)
+QTransform KoGamutMask::maskToViewTransform(qreal viewSize)
 {
     // apply mask rotation before drawing
     QPointF centerPoint(viewSize*0.5, viewSize*0.5);
@@ -198,7 +198,7 @@ QTransform KoGamutMask::maskToViewTransform(quint8 viewSize)
     return transform;
 }
 
-QTransform KoGamutMask::viewToMaskTransform(quint8 viewSize)
+QTransform KoGamutMask::viewToMaskTransform(qreal viewSize)
 {
     QPointF centerPoint(viewSize*0.5, viewSize*0.5);
